backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
firebase-admin
pandas
numpy
scikit-learn
joblib
python-multipart
